+++ src/pytask_parallel/config.py
@@ -3,6 +3,7 @@ import os
 
 from _pytask.config import hookimpl
 from _pytask.shared import get_first_non_none_value
+from pytask_parallel.backends import PARALLEL_BACKENDS_DEFAULT
 from pytask_parallel.callbacks import delay_callback
 from pytask_parallel.callbacks import n_workers_callback
 from pytask_parallel.callbacks import parallel_backend_callback
@@ -33,7 +34,7 @@ def pytask_parse_config(config, config_from_cli, config_from_file):
         config_from_cli,
         config_from_file,
         key="parallel_backend",
-        default="processes",
+        default=PARALLEL_BACKENDS_DEFAULT,
         callback=parallel_backend_callback,
     )