+++ src/pytask_parallel/plugin.py
@@ -3,6 +3,7 @@ from _pytask.config import hookimpl
 from pytask_parallel import build
 from pytask_parallel import config
 from pytask_parallel import execute
+from pytask_parallel import logging
 
 
 @hookimpl
@@ -11,3 +12,4 @@ def pytask_add_hooks(pm):
     pm.register(build)
     pm.register(config)
     pm.register(execute)
+    pm.register(logging)