+++ CHANGES.rst
@@ -3,7 +3,14 @@ Changes
 
 This is a record of all past pytask-parallel releases and what went into them in reverse
 chronological order. Releases follow `semantic versioning <https://semver.org/>`_ and
-all releases are available on `Anaconda.org <https://anaconda.org/pytask/pytask-parallel>`_.
+all releases are available on `Anaconda.org
+<https://anaconda.org/pytask/pytask-parallel>`_.
+
+
+0.0.4 - 2020-10-30
+------------------
+
+- :gh:`4` implement an executor with ``loky``.
 
 
 0.0.3 - 2020-09-12